requirements.txt
flask==3.0.3
flask-socketio==5.3.6
eventlet==0.36.1
sqlmodel==0.0.22
paho-mqtt==2.1.0
python-dotenv==1.0.1
